Created the battle engine's entry and result structs Started the pokemon state. Added a timer for main. serebii_pokemon_data_source now has a separate method for fetching moves. WormadamS and WormadamT now are bug (not grass) BattleFactoryPokemon now stores Move enums rather than MoveInfo pointers. Effort values are now in an array instead of a map. Added some notes to ability.h IVs are now 8 bit.

+++ stat.h
@@ -30,7 +30,7 @@ static const std::unordered_map<std::string, Stat> STRING_TO_STAT_MAP{
     {"speed", Stat::SPEED}
 };
 
-inline int get_iv_for_battle_factory(const int round_number) {
+inline uint8_t get_iv_for_battle_factory(const int round_number) {
     if (round_number == 0) return 0;
     const int iv = (round_number - 1) * 4;
     return (round_number < 8) ? iv : 31;
